Version 1.1.0 release
- U++ framework code update
- C++ performance and bug correction

+++ app/PxView/lib/px_head.c
@@ -3,6 +3,7 @@
 #endif
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <fcntl.h>
 #include <time.h>
@@ -287,8 +288,8 @@ int put_px_head(pxdoc_t *pxdoc, pxhead_t *pxh, pxstream_t *pxs) {
 			 * In several files it was just 12 in .DB and 17 in .PX files. */
 			put_short_le((char *)&pxhead.unknown12x13, 12);
 			put_short_le((char *)&pxhead.primaryKeyFields, pxh->px_primarykeyfields);
-			put_long_le((char *)&pxhead.primaryIndexWorkspace, (int) (basehead-100));  /* just to set a value */
-			put_long_le((char *)&pxhead.unknownPtr1A, (int) (basehead-500));  /* just to set a value */
+			put_long_le((char *)&pxhead.primaryIndexWorkspace, (intptr_t) (basehead-100));  /* just to set a value */
+			put_long_le((char *)&pxhead.unknownPtr1A, (intptr_t) (basehead-500));  /* just to set a value */
 			break;
 		case pxfFileTypPrimIndex:
 			put_short_le((char *)&pxhead.unknown12x13, 17);
@@ -321,8 +322,8 @@ int put_px_head(pxdoc_t *pxdoc, pxhead_t *pxh, pxstream_t *pxs) {
 		dataheadoffset = 0x58;
 	}
 	/* All the pointers, though we probably don't need them for a valid file */
-	put_long_le((char *)&pxhead.fldInfoPtr, (int) (basehead+dataheadoffset));
-	put_long_le((char *)&pxhead.tableNamePtrPtr, (int) (basehead+dataheadoffset+pxh->px_numfields*2));
+	put_long_le((char *)&pxhead.fldInfoPtr, (intptr_t) (basehead+dataheadoffset));
+	put_long_le((char *)&pxhead.tableNamePtrPtr, (intptr_t) (basehead+dataheadoffset+pxh->px_numfields*2));
 	switch(pxh->px_fileversion) {
 		case 70:
 			pxhead.fileVersionID = 0x0C;
@@ -433,7 +434,7 @@ int put_px_head(pxdoc_t *pxdoc, pxhead_t *pxh, pxstream_t *pxs) {
 	 * with numfields fields specifications (each 2 Bytes), followed
 	 * by this pointer (tableNamePtr) and numfield pointers to the
 	 * field names. */
-	put_long_le((char *)&ptr, (int) (basehead+dataheadoffset+pxh->px_numfields*(2+4)+4));
+	put_long_le((char *)&ptr, (intptr_t) (basehead+dataheadoffset+pxh->px_numfields*(2+4)+4));
 	if(pxdoc->write(pxdoc, pxs, 4, &ptr) < 1) {
 		px_error(pxdoc, PX_RuntimeError, _("Could not write pointer to tablename."));
 		return -1;
@@ -443,7 +444,7 @@ int put_px_head(pxdoc_t *pxdoc, pxhead_t *pxh, pxstream_t *pxs) {
 	 * numfields sizeof(* Fieldname) + sizeof(* Tablename) + strlen(tablename)
 	 */
 	if(!isindex) {
-		base = (int) basehead+dataheadoffset+pxh->px_numfields*(2+4)+4+tablenamelen;
+		base = (intptr_t) basehead+dataheadoffset+pxh->px_numfields*(2+4)+4+tablenamelen;
 		pxf = pxh->px_fields;
 		offset = 0;
 		for(i=0; i<pxh->px_numfields; i++, pxf++) {
